LPS-131949 Migrate Forms to use the new Data Engine App
- Ticket: https://issues.liferay.com/browse/LPS-131949
