+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/admin/js/config/initialConfigState.es.js
@@ -13,8 +13,31 @@
  */
 
 import {INITIAL_CONFIG_STATE as INITIAL_FORM_RENDERER_CONFIG_STATE} from 'data-engine-js-components-web/js/core/config/index.es';
+import React from 'react';
+
+import ElementSetList from '../plugins/field-sidebar/components/ElementSetList.es';
+
+const tabs = [
+	{
+		label: Liferay.Language.get('element-sets'),
+		render: ({searchTerm}) => <ElementSetList searchTerm={searchTerm} />,
+	},
+];
 
 export const INITIAL_CONFIG_STATE = {
 	...INITIAL_FORM_RENDERER_CONFIG_STATE,
-	allowNestedFields: true,
+	allowFieldSets: false,
+	allowMultiplePages: false,
+	allowNestedFields: false,
+	allowRules: true,
+	allowSuccessPage: false,
+	disabledProperties: [],
+	disabledTabs: [],
+	tabs,
+	unimplementedProperties: [
+		'fieldNamespace',
+		'readOnly',
+		'visibilityExpression',
+	],
+	visibleProperties: [],
 };
