+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/admin/js/eventTypes.es.js
@@ -20,7 +20,6 @@ const RULE = {
 };
 
 const SIDEBAR = {
-	BLUR: 'sidebar_blur',
 	CHANGE_FIELD_TYPE: 'sidebar_change_field_type',
 	CHANGES_CANCEL: 'sidebar_changes_cancel',
 };
