+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/admin/js/plugins/field-sidebar/components/FormsFieldSidebar.es.js
@@ -19,72 +19,23 @@ import {FieldsSidebar} from 'data-engine-taglib';
 import React from 'react';
 
 import {EVENT_TYPES} from '../../../eventTypes.es';
-import ElementSetList from './ElementSetList.es';
 
 export const FormsFieldSidebar = ({title}) => {
-	const {
-		dataProviderInstanceParameterSettingsURL,
-		dataProviderInstancesURL,
-		fieldSetDefinitionURL,
-		fieldTypes,
-		functionsMetadata,
-		functionsURL,
-		portletNamespace,
-	} = useConfig();
+	const {fieldTypes} = useConfig();
 	const {
 		activePage,
 		defaultLanguageId,
 		editingLanguageId,
-		fieldSets,
 		focusedField,
 		pages,
 		rules,
 	} = useFormState();
 
-	const tabs = [
-		{
-			label: Liferay.Language.get('element-sets'),
-			render: ({searchTerm}) => (
-				<ElementSetList
-					definitionURL={fieldSetDefinitionURL}
-					editingLanguageId={editingLanguageId}
-					elementSets={fieldSets}
-					namespace={portletNamespace}
-					searchTerm={searchTerm}
-				/>
-			),
-		},
-	];
-
 	const dispatch = useForm();
-
-	const config = {
-		allowFieldSets: false,
-		allowMultiplePages: false,
-		allowNestedFields: false,
-		allowRules: true,
-		allowSuccessPage: false,
-		disabledProperties: [],
-		disabledTabs: [],
-		ruleSettings: {
-			dataProviderInstanceParameterSettingsURL,
-			dataProviderInstancesURL,
-			functionsMetadata,
-			functionsURL,
-		},
-		tabs,
-		unimplementedProperties: [
-			'fieldNamespace',
-			'readOnly',
-			'visibilityExpression',
-		],
-		visibleProperties: [],
-	};
 
 	return (
 		<FieldsSidebar
 			classNames={classNames}
-			config={config}
 			dataLayout={{
 				dataLayoutFields: [],
 				dataLayoutPages: [],
