+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/resources/META-INF/resources/admin/js/reducers/sidebarReducer.es.js
@@ -13,6 +13,7 @@
  */
 
 import {PagesVisitor} from 'data-engine-js-components-web';
+import {EVENT_TYPES as CORE_EVENT_TYPES} from 'data-engine-taglib';
 import {
 	FieldUtil,
 	RulesSupport,
@@ -27,7 +28,7 @@ import {EVENT_TYPES} from '../eventTypes.es';
  */
 export default (state, action) => {
 	switch (action.type) {
-		case EVENT_TYPES.SIDEBAR.BLUR: {
+		case CORE_EVENT_TYPES.SIDEBAR.FIELD.BLUR: {
 			const {focusedField} = state;
 
 			if (
